+++ lib/Metrics.ts
@@ -1,9 +1,8 @@
-import * as Debug from "debug";
 import * as promClient from "prom-client";
-
-const debug = Debug("roach:metrics");
 const promDefaultMetrics = promClient.collectDefaultMetrics;
 const promRegistry = promClient.Registry;
+
+const UNDERSCORE_REGEX = /-/g;
 
 export class Metrics {
 
@@ -33,6 +32,19 @@ export class Metrics {
         return this.register;
     }
 
+    private cleanMetricName(name: string): string {
+        return name.replace(UNDERSCORE_REGEX, "_");
+    }
+
+    private cleanLabels(labels: { [labelName: string]: string }): { [labelName: string]: string } {
+
+        Object.keys(labels).forEach((labelName) => {
+            labels[labelName] = this.cleanMetricName(labels[labelName]);
+        });
+
+        return labels;
+    }
+
     private getCounter(key: string): promClient.Counter {
 
         if (this.metrics[key]) {
@@ -43,7 +55,7 @@ export class Metrics {
             name: `${key}`,
             help: `${key}_help`,
             registers: [this.register],
-            labelNames: ["na", "prefix"],
+            // labelNames: [],
         });
 
         return this.metrics[key];
@@ -62,51 +74,49 @@ export class Metrics {
             name: `${key}`,
             help: `${key}_help`,
             registers: [this.register],
-            labelNames: ["na", "prefix"], // Gauges require fixed prefixes
+            // labelNames: [], // Gauges require fixed prefixes
         });
 
         return this.metrics[key] as promClient.Gauge;
     }
 
-    public inc(key: string, val: number = 1) {
+    public inc(key: string, val: number = 1, labels: { [labelName: string]: string } = {}) {
 
+        key = this.cleanMetricName(key);
         const prefix = this.prefix;
         const fullKey = prefix ? `${prefix}_${key}` : key;
         const counter = this.getCounter(fullKey);
 
         counter.inc(
-            { prefix },
+            this.cleanLabels(labels),
             val,
             Date.now(),
         );
-
     }
 
-    public set(key: string, val: number) {
+    public set(key: string, val: number, labels: { [labelName: string]: string } = {}) {
 
         if (val === null || val === undefined) {
             throw new Error(`Please provide value on set ${key}`);
         }
 
+        key = this.cleanMetricName(key);
         const prefix = this.prefix;
         const fullKey = prefix ? `${prefix}_${key}` : key;
         const gauge = this.getGauge(fullKey);
 
         gauge.set(
-            { prefix },
+            this.cleanLabels(labels),
             val,
             Date.now(),
         );
     }
 
     public registerDefault() {
-
         this.defaultMetricsIntv = promDefaultMetrics({
             register: this.register,
             timeout: 5000,
         });
-
-        debug("default metrics active.");
     }
 
     public close() {