+++ lib/PubSubHandler.ts
@@ -2,7 +2,6 @@ import { PubSub, Subscription } from "@google-cloud/pubsub";
 
 import { RoachConfig } from "./interfaces";
 import RoachStorm from "./RoachStorm";
-import MessageHandler from "./MessageHandler";
 
 export default class PubSubHandler {
 
@@ -20,7 +19,7 @@ export default class PubSubHandler {
             .topic(topic)
             .publish(Buffer.from(message));
 
-        this.roachStorm.metrics.inc(`pubsub_msg_out_${(MessageHandler.cleanTopicNameForMetrics(topic))}`);
+        this.roachStorm.metrics.inc(`pubsub_msg_out`, 1, { topic });
         return messageId;
     }